+++ api/src/utils/slates.ts
@@ -8,8 +8,9 @@ import { getContracts } from './eth';
 import * as config from './config';
 import { nonEmptyString } from './validation';
 import { getProposalsForRequests } from './requests';
+import { mapProposalsToRequests } from './proposals';
 
-const { IpfsMetadata, Slate, Request } = require('../models');
+const { IpfsMetadata, Slate } = require('../models');
 const { toUtf8String, bigNumberify: BN, getAddress } = ethers.utils;
 const { tokenCapacitorAddress } = config.contracts;
 
@@ -156,27 +157,7 @@ async function getSlateWithMetadata(slateID, slate, metadataHash, incumbent, req
     // console.log('proposalMultihashes:', proposalMultihashes);
     // console.log('');
 
-    const proposalsWithIds = await Promise.all(
-      proposals.map(async (proposal, i) => {
-        const multihash = Buffer.from(proposalMultihashes[i]).toString('hex');
-        const request = await Request.findOne(
-          {
-            where: {
-              metadataHash: `0x${multihash}`,
-            },
-          },
-          { raw: true }
-        );
-        if (request == null) {
-          return proposal;
-        }
-        return {
-          ...proposal,
-          proposalID: request.proposalID,
-          requestID: request.requestID,
-        };
-      })
-    );
+    const proposalsWithIds = await mapProposalsToRequests(proposals, proposalMultihashes);
 
     // --------------------------
     // COMBINE/RETURN SLATE DATA
